requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pygame>=2.5.0
PyOpenGL>=3.1.6
PyYAML>=6.0
Pillow>=9.0.0
